Theme is has changed! There are multiple changes

+++ app/src/main/java/com/example/liderstoreagent/data/models/clientmodel/ClientsData.kt
@@ -2,5 +2,5 @@ package com.example.liderstoreagent.data.models.clientmodel
 
 data class ClientsData(
     val client: Client,
-    val total_debt: Int
+    val total_debt: Double
 )
